+++ R/xportr-package.R
@@ -110,12 +110,14 @@
 #' @importFrom utils capture.output str tail packageVersion
 #' @importFrom stringr str_detect str_extract str_replace str_replace_all
 #' @importFrom readr parse_number
-#' @importFrom purrr map_chr map2_chr walk walk2 map map_dbl pluck
+#' @importFrom purrr map_chr map2_chr walk iwalk map map_dbl pluck
 #' @importFrom janitor make_clean_names
 #' @importFrom tm stemDocument
 #' @importFrom graphics stem
 #' @importFrom magrittr %>% extract2
-#'
+#' @importFrom checkmate assert assert_character assert_choice assert_data_frame
+#' assert_integer assert_logical assert_string makeAssertion check_data_frame
+#' check_r6 test_data_frame test_string vname
 "_PACKAGE"
 
 globalVariables(c(